requirements.txt
fastapi
uvicorn==0.15.0
pydantic==1.8.2
jinja2==3.0.1
requests==2.26.0
pillow==8.3.1
#torch==1.9.0
#torchvision==0.10.0
#layoutparser==0.3.4
#detectron2==0.6